Standardize visualize utilities
